+++ projects/cm-d3-graphs/src/lib/services/d3-utility.service.ts
@@ -273,11 +273,13 @@ export class D3UtilityService {
       collapsedColor: config.nodes.collapsedColor,
       expandedColor: config.nodes.expandedColor,
       strokeColor: config.nodes.strokeColor,
+      strokeWidth: config.nodes.strokeWidth,
       labelColor: config.label.color,
       circleRadius: config.nodes.circleRadius,
       rectangleDimensions: config.nodes.rectangleDimensions,
       squareDimensions: config.nodes.squareDimensions,
-      rhombusDimensions: config.nodes.rhombusDimensions
+      rhombusDimensions: config.nodes.rhombusDimensions,
+      additionalClasses: config.nodes.additionalClasses
     }, inputData.node ? inputData.node : {});
     newData.link = D3UtilityService.merge({
       color: 'lightsteelblue',
@@ -291,12 +293,14 @@ export class D3UtilityService {
       collapsedColor: config.nodes.collapsedColor,
       expandedColor: config.nodes.expandedColor,
       strokeColor: config.nodes.strokeColor,
+      strokeWidth: config.nodes.strokeWidth,
       labelColor: config.label.color,
       icon: config.nodes.icon,
       circleRadius: config.nodes.circleRadius,
       rectangleDimensions: config.nodes.rectangleDimensions,
       squareDimensions: config.nodes.squareDimensions,
-      rhombusDimensions: config.nodes.rhombusDimensions
+      rhombusDimensions: config.nodes.rhombusDimensions,
+      additionalClasses: config.nodes.additionalClasses
     }, inputData.node ? inputData.node : {});
     newData.link = D3UtilityService.merge({
       color: 'lightsteelblue',